move ingress regression tests into regression suite
