--- projects/clusteringress/pkg/status/status_syncer_suite_test.go
@@ -1,15 +1,0 @@
-package status_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	"github.com/onsi/ginkgo/reporters"
-	. "github.com/onsi/gomega"
-)
-
-func TestStatusSyncer(t *testing.T) {
-	RegisterFailHandler(Fail)
-	junitReporter := reporters.NewJUnitReporter("junit.xml")
-	RunSpecsWithDefaultAndCustomReporters(t, "StatusSyncer Suite", []Reporter{junitReporter})
-}
