--- projects/ingress/pkg/status/status_suite_test.go
@@ -1,16 +1,0 @@
-package status_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	"github.com/onsi/ginkgo/reporters"
-	. "github.com/onsi/gomega"
-	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
-)
-
-func TestStatus(t *testing.T) {
-	RegisterFailHandler(Fail)
-	junitReporter := reporters.NewJUnitReporter("junit.xml")
-	RunSpecsWithDefaultAndCustomReporters(t, "Status Suite", []Reporter{junitReporter})
-}
